+++ crates/bytecode/src/chunk.rs
@@ -1,10 +1,7 @@
 use crate::InstructionReader;
 use koto_memory::Ptr;
-use koto_parser::{ConstantPool, Span};
-use std::{
-    fmt::{self, Write},
-    path::{Path, PathBuf},
-};
+use koto_parser::{ConstantPool, KString, Span};
+use std::fmt::{self, Write};
 
 /// Debug information for a Koto program
 #[derive(Clone, Debug, Default, PartialEq, Eq)]
@@ -56,27 +53,12 @@ pub struct Chunk {
     /// The constant data associated with the chunk's bytecode
     pub constants: ConstantPool,
     /// The path of the program's source file
-    pub source_path: Option<PathBuf>,
+    pub source_path: Option<KString>,
     /// Debug information associated with the chunk's bytecode
     pub debug_info: DebugInfo,
 }
 
 impl Chunk {
-    /// Initializes a Chunk
-    pub fn new(
-        bytes: Box<[u8]>,
-        constants: ConstantPool,
-        source_path: Option<&Path>,
-        debug_info: DebugInfo,
-    ) -> Self {
-        Self {
-            bytes,
-            constants,
-            source_path: source_path.map(Path::to_path_buf),
-            debug_info,
-        }
-    }
-
     /// Returns a [String] displaying the instructions contained in the compiled [Chunk]
     pub fn bytes_as_string(chunk: &Chunk) -> String {
         let mut iter = chunk.bytes.iter();