+++ crates/koto/examples/poetry/src/main.rs
@@ -7,13 +7,9 @@ use hotwatch::{
     notify::event::ModifyKind,
     Event, EventKind,
 };
-use koto::{Koto, KotoSettings};
+use koto::prelude::*;
 use poetry::*;
-use std::{
-    fs,
-    path::{Path, PathBuf},
-    time::Duration,
-};
+use std::{fs, path::PathBuf, time::Duration};
 
 fn version_string() -> String {
     format!("{} {}", env!("CARGO_PKG_NAME"), env!("CARGO_PKG_VERSION"))
@@ -92,9 +88,7 @@ fn main() -> Result<()> {
         .insert("poetry", koto_bindings::make_module());
     koto.prelude().insert("random", koto_random::make_module());
 
-    let script_path = PathBuf::from(args.script);
-    koto.set_script_path(Some(&script_path))
-        .expect("Failed to set script path");
+    let script_path = KString::from(args.script);
 
     if args.watch {
         if let Err(e) = compile_and_run(&mut koto, &script_path) {
@@ -104,16 +98,19 @@ fn main() -> Result<()> {
         let mut hotwatch = Hotwatch::new_with_custom_delay(Duration::from_secs_f64(0.25))
             .context("Failed to initialize file watcher")?;
         hotwatch
-            .watch(script_path.clone(), move |event: Event| {
-                match event.kind {
-                    EventKind::Create(_) | EventKind::Modify(ModifyKind::Data(_)) => {
-                        if let Err(error) = compile_and_run(&mut koto, &script_path) {
-                            eprintln!("{error}");
+            .watch(&script_path, {
+                let script_path = script_path.clone();
+                move |event: Event| {
+                    match event.kind {
+                        EventKind::Create(_) | EventKind::Modify(ModifyKind::Data(_)) => {
+                            if let Err(error) = compile_and_run(&mut koto, &script_path) {
+                                eprintln!("{error}");
+                            }
                         }
+                        _ => {}
                     }
-                    _ => {}
+                    Flow::Continue
                 }
-                Flow::Continue
             })
             .context("Failed to watch file!")?;
         hotwatch.run();
@@ -123,10 +120,14 @@ fn main() -> Result<()> {
     }
 }
 
-fn compile_and_run(koto: &mut Koto, script_path: &Path) -> Result<()> {
-    let script = fs::read_to_string(script_path)?;
-    koto.compile(&script)
-        .context("Error while compiling script")?;
+fn compile_and_run(koto: &mut Koto, script_path: &KString) -> Result<()> {
+    let script = fs::read_to_string(script_path.as_str())?;
+    koto.compile(CompileArgs {
+        script: &script,
+        script_path: Some(script_path.clone()),
+        compiler_settings: Default::default(),
+    })
+    .context("Error while compiling script")?;
     koto.run().context("Error while running script")?;
     Ok(())
 }